+++ src/Repository/RepositoryAdapterAwareInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace ConductorCore\Repository;
+
+interface RepositoryAdapterAwareInterface
+{
+    public function setRepositoryAdapter(RepositoryAdapterInterface $repositoryAdapter): void;
+}